Simplified injectable config translators and named them to avoid clashes
