+++ Dockerfile
@@ -22,7 +22,7 @@ RUN \
   chmod 0777 -R /var/lib/mysql /var/log/mysql && \
   chmod 0775 -R /etc/mysql && \
   chmod 0755 -R /hooks && \
-  cd /opt/configurability/src/ && \
+  cd /opt/configurability/src/mariadb_config_translator && \
   pip --no-cache install --upgrade pip && \
   pip --no-cache install --upgrade .
