+++ files/etc/configurability/my.ini
@@ -3,3 +3,4 @@ enabled = true
 module = configurability.ini_merge_process
 ini_file_path = /etc/mysql/mariadb.conf.d/50-server.cnf
 configuration_file_name = configuration-mysql.json
+config_translator_module = mariadb_config_translator
