--- files/opt/configurability/src/injectable/config/__init__.py
@@ -1,6 +1,0 @@
-import sql_mode as sql_mode
-
-def process(key, value):
-    if key == 'sql_mode':
-        return sql_mode.process(value)
-    return value
