--- files/opt/configurability/src/injectable/config/sql_mode.py
@@ -1,4 +1,0 @@
-def process(value):
-	if isinstance(value, list):
-		return ','.join(value)
-	return value
